requirements.txt
streamlit==1.29.0
pillow==10.1.0
requests==2.31.0
openai>=1.0.0
dashscope>=1.25.0